feat(web): configure static export build for Go binary embedding
- Add output: "export" to next.config.ts for static HTML/CSS/JS generation
- Remove force-dynamic directive from 13 page files (legacy server-side auth)
- Split 4 dynamic [id] routes into server page.tsx + client component pairs
  (Next.js 16 Turbopack rejects generateStaticParams in "use client" files)
- Use placeholder params in generateStaticParams to satisfy Cache Components
  validation (Next.js 16 rejects empty arrays with misleading error message)
- Update start script from "next start" to "npx serve out"
- Build produces out/ directory: 14 static + 4 SSG pages in ~6 seconds

Made-with: Cursor
