+++ _projects/2026-03/20260314.03.web-console-oss-migration/checkpoints/2026-03-14-session-4.md
@@ -1,0 +1,76 @@
+# Session Notes: 2026-03-14 (Session 4)
+
+## Accomplishments
+
+- **T04 Complete**: Configure Static Export Build
+  - Web console builds to `out/` directory as static HTML/CSS/JS
+  - All 18 routes generate successfully (14 static, 4 SSG with dynamic params)
+  - Build completes in ~6 seconds
+  - Static server test confirms all routes return 200
+
+## Files Created
+
+- `src/app/agents/[id]/AgentDetailPage.tsx` — Client component extracted from page.tsx
+- `src/app/mcp-servers/[id]/McpServerDetailPage.tsx` — Client component extracted from page.tsx
+- `src/app/sessions/[id]/SessionDetailPage.tsx` — Client component extracted from page.tsx
+- `src/app/skills/[id]/SkillDetailPage.tsx` — Client component extracted from page.tsx
+
+## Files Modified
+
+- `next.config.ts` — Added `output: "export"`
+- `package.json` — Changed `start` script from `next start` to `npx serve out`
+- 13 page files — Removed `export const dynamic = "force-dynamic"` (legacy from stigmer-cloud server-side auth)
+- 4 dynamic route `page.tsx` files — Rewritten as server component wrappers with `generateStaticParams`
+
+## Decisions Made
+
+- **Server/client split for dynamic routes**: Next.js 16 (Turbopack) explicitly rejects `generateStaticParams` in `"use client"` files. Each dynamic route was split into a server `page.tsx` (exports `generateStaticParams`, renders client component) and a colocated client component file (has `"use client"`, uses `useParams()`).
+- **Placeholder params instead of empty arrays**: Next.js 16's Cache Components feature (enabled by default) rejects `generateStaticParams` returning `[]`. Each route returns `[{ id: "__placeholder__" }]` instead. This generates a pre-rendered page at `/<route>/__placeholder__` which is harmless — the client component will attempt an API fetch for a non-existent ID and show its error state.
+- **Runtime config deferred to T05**: Build-time `NEXT_PUBLIC_*` env vars are sufficient for both local and cloud builds. Runtime config will be implemented alongside Go embedding.
+- **Dockerfile not touched**: Still references Yarn and standalone mode — will be fully rewritten in T05/T07.
+
+## Surprises Discovered
+
+- **`generateStaticParams` + `"use client"` incompatibility**: The plan anticipated this as a risk and prepared a mitigation (server/client split). Next.js 16 with Turbopack rejects the combination at compile time: "App pages cannot use both 'use client' and export function 'generateStaticParams()'."
+- **Empty `generateStaticParams` array rejected**: This was NOT anticipated. Next.js 16's Cache Components (enabled by default) requires `generateStaticParams` to return at least one param set for build-time validation. The error message is misleading — it says "missing generateStaticParams()" instead of indicating the empty array is the problem. This required iterative debugging to identify the root cause.
+
+## Learnings
+
+- Next.js 16 has Cache Components enabled by default. This changes build-time validation behavior for `generateStaticParams` — empty arrays are no longer valid, even when you genuinely have no build-time-known IDs.
+- The error message "Page is missing generateStaticParams()" can mean either (a) the function is truly not exported, or (b) it returns an empty array with Cache Components enabled. The misleading error message cost significant debugging time.
+- The `__placeholder__` pattern from the Next.js docs is the official workaround for this limitation.
+- `npx serve` handles clean URLs automatically (redirects `/agents.html` to `/agents`), making it suitable for quick local testing of static exports.
+
+## Build Output Verification
+
+```
+Route (app)
+├ ○ /                     (Static)
+├ ○ /agents               (Static)
+├ ● /agents/[id]          (SSG — /agents/__placeholder__)
+├ ○ /catalog              (Static)
+├ ○ /draft/agent          (Static)
+├ ○ /draft/mcp-server     (Static)
+├ ○ /draft/skill          (Static)
+├ ○ /mcp-servers          (Static)
+├ ● /mcp-servers/[id]     (SSG — /mcp-servers/__placeholder__)
+├ ○ /run                  (Static)
+├ ○ /sessions             (Static)
+├ ● /sessions/[id]        (SSG — /sessions/__placeholder__)
+├ ○ /skills               (Static)
+└ ● /skills/[id]          (SSG — /skills/__placeholder__)
+```
+
+## Open Questions
+
+- The `__placeholder__` pre-rendered pages exist in the `out/` directory. The Go server in T05 could optionally exclude them from serving, or simply let them render naturally (the client component will show a "not found" or error state for the non-existent ID).
+- SPA fallback routing in T05: when a user navigates directly to `/agents/some-real-id`, the Go server must serve the `__placeholder__` HTML (or a generic shell) since no page is pre-rendered for that specific ID.
+- `trailingSlash` configuration: deferred to T05 based on Go embedding requirements.
+
+## Next Session Plan
+
+- Start T05: Embed Web UI in stigmer-server
+  - `//go:embed` the `out/` directory into the Go binary
+  - HTTP handler to serve static files
+  - SPA fallback routing for dynamic `[id]` routes
+  - Runtime config endpoint if needed
