+++ client-apps/web/src/app/draft/agent/page.tsx
@@ -3,8 +3,6 @@
 import { DraftPage } from "@/components/draft/DraftPage";
 import { AGENT_DRAFT_CONFIG } from "@/config/draft";
 
-export const dynamic = "force-dynamic";
-
 export default function DraftAgentPage() {
   return <DraftPage config={AGENT_DRAFT_CONFIG} />;
 }
