+++ client-apps/web/src/app/draft/mcp-server/page.tsx
@@ -3,8 +3,6 @@
 import { DraftPage } from "@/components/draft/DraftPage";
 import { MCP_SERVER_DRAFT_CONFIG } from "@/config/draft";
 
-export const dynamic = "force-dynamic";
-
 export default function DraftMcpServerPage() {
   return <DraftPage config={MCP_SERVER_DRAFT_CONFIG} />;
 }
